♻️ decompose /trade reply modules and unify variant paths
Split deals/variants/tables out of tradeReply, always group by variant, and parallelize icon renders.
